basic logging for the login
