+++ openbb_terminal/core/log/generation/common.py
@@ -1,0 +1,19 @@
+# IMPORTATION STANDARD
+import logging
+
+# IMPORTATION THIRDPARTY
+
+
+# IMPORTATION INTERNAL
+from openbb_terminal.core.log.generation.path_tracking_file_handler import (
+    PathTrackingFileHandler,
+)
+
+
+def do_rollover():
+    """RollOver the log file."""
+
+    for handler in logging.getLogger().handlers:
+        print(handler)
+        if isinstance(handler, PathTrackingFileHandler):
+            handler.doRollover()
