+++ openbb_terminal/core/log/generation/settings_logger.py
@@ -11,9 +11,7 @@ from types import FunctionType, ModuleType
 import openbb_terminal.feature_flags as obbff
 from openbb_terminal import config_terminal as cfg
 from openbb_terminal.terminal_helper import is_packaged_application
-from openbb_terminal.core.log.generation.path_tracking_file_handler import (
-    PathTrackingFileHandler,
-)
+from openbb_terminal.core.log.generation.common import do_rollover
 
 
 SENSITIVE_WORDS = [
@@ -31,14 +29,15 @@ SENSITIVE_WORDS = [
 logger = logging.getLogger(__name__)
 
 
-def log_all_settings() -> None:
+def log_all_settings(with_rollover: bool = True) -> None:
     """Log all settings"""
     log_settings()
     log_config_terminal()
     log_feature_flags()
     log_keys()
 
-    do_rollover()
+    if with_rollover:
+        do_rollover()
 
 
 def log_settings() -> None:
@@ -117,11 +116,3 @@ def log_keys() -> None:
             current_keys[cfg_var_name] = "not_defined"
 
     logger.info("KEYS: %s ", json.dumps(current_keys))
-
-
-def do_rollover():
-    """RollOver the log file."""
-
-    for handler in logging.getLogger().handlers:
-        if isinstance(handler, PathTrackingFileHandler):
-            handler.doRollover()
